# contracts/web-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "One Person Lab Web API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register a public Web account.",
        "responses": {
          "201": {
            "description": "Registered and session cookie set."
          },
          "400": {
            "description": "Invalid JSON or credential input."
          },
          "409": {
            "description": "Email is already registered."
          },
          "405": {
            "description": "Method not allowed."
          },
          "500": {
            "description": "Account could not be created."
          },
          "503": {
            "description": "Session secret is missing."
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login and set an HttpOnly session cookie.",
        "responses": {
          "200": {
            "description": "Authenticated."
          },
          "400": {
            "description": "Invalid JSON."
          },
          "401": {
            "description": "Invalid credentials."
          },
          "405": {
            "description": "Method not allowed."
          },
          "503": {
            "description": "Session secret is missing."
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Clear the current session.",
        "responses": {
          "204": {
            "description": "Logged out."
          },
          "405": {
            "description": "Method not allowed."
          }
        }
      }
    },
    "/api/session/current": {
      "get": {
        "summary": "Return current user session state.",
        "responses": {
          "200": {
            "description": "Current authenticated session."
          },
          "401": {
            "description": "Authentication required."
          },
          "405": {
            "description": "Method not allowed."
          }
        }
      }
    },
    "/api/settings/model-provider": {
      "get": {
        "summary": "Return fixed provider state without raw API key.",
        "responses": {
          "200": {
            "description": "Provider state with fixed baseUrl and masked key."
          },
          "401": {
            "description": "Authentication required."
          },
          "405": {
            "description": "Method not allowed."
          }
        }
      },
      "put": {
        "summary": "Save user API key for the fixed provider.",
        "responses": {
          "200": {
            "description": "Provider key saved without returning raw key."
          },
          "400": {
            "description": "Invalid JSON or API key."
          },
          "401": {
            "description": "Authentication required."
          },
          "405": {
            "description": "Method not allowed."
          },
          "500": {
            "description": "Provider key could not be saved."
          },
          "503": {
            "description": "API key encryption secret is missing."
          }
        }
      }
    },
    "/api/chat": {
      "post": {
        "summary": "Send ordinary chat through the fixed provider or return runtime gate.",
        "responses": {
          "200": {
            "description": "Assistant response."
          },
          "400": {
            "description": "Invalid JSON, message, or missing provider key."
          },
          "401": {
            "description": "Authentication required."
          },
          "404": {
            "description": "Conversation not found for current user."
          },
          "405": {
            "description": "Method not allowed."
          },
          "409": {
            "description": "Runtime required for OPL capability."
          },
          "429": {
            "description": "Quota exceeded before upstream call."
          },
          "502": {
            "description": "Upstream chat failed."
          },
          "503": {
            "description": "Provider key binding cannot be decrypted."
          },
          "504": {
            "description": "Upstream chat timed out."
          }
        }
      }
    },
    "/api/chat/conversations": {
      "get": {
        "summary": "List current user's conversations.",
        "responses": {
          "200": {
            "description": "Conversation list."
          },
          "401": {
            "description": "Authentication required."
          },
          "405": {
            "description": "Method not allowed."
          }
        }
      }
    },
    "/api/chat/conversations/{conversationId}": {
      "get": {
        "summary": "Read one current-user conversation.",
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation detail."
          },
          "400": {
            "description": "Invalid conversation id."
          },
          "401": {
            "description": "Authentication required."
          },
          "404": {
            "description": "Conversation not found for current user."
          },
          "405": {
            "description": "Method not allowed."
          }
        }
      }
    },
    "/api/account/audit-events": {
      "get": {
        "summary": "List current user's sanitized audit events.",
        "responses": {
          "200": {
            "description": "Sanitized user audit events."
          },
          "401": {
            "description": "Authentication required."
          },
          "405": {
            "description": "Method not allowed."
          }
        }
      }
    },
    "/api/account/billing-summary": {
      "get": {
        "summary": "Read current user's sanitized billing, quota, and audit projection.",
        "responses": {
          "200": {
            "description": "Sanitized billing handoff and local quota/audit summary."
          },
          "401": {
            "description": "Authentication required."
          },
          "405": {
            "description": "Method not allowed."
          }
        }
      }
    },
    "/api/medopl/runtime/status": {
      "get": {
        "summary": "Read sanitized MedOPL runtime status projection.",
        "responses": {
          "200": {
            "description": "Readonly MedOPL runtime status projection."
          },
          "405": {
            "description": "Method not allowed."
          }
        }
      }
    },
    "/api/medopl/materials-deliverables/projection": {
      "get": {
        "summary": "Read sanitized MedOPL materials and deliverables projection.",
        "responses": {
          "200": {
            "description": "Readonly MedOPL materials and deliverables projection."
          },
          "405": {
            "description": "Method not allowed."
          }
        }
      }
    },
    "/api/opl/snapshot": {
      "get": {
        "summary": "Read OPL readonly snapshot projection.",
        "responses": {
          "200": {
            "description": "Readonly snapshot projection."
          },
          "405": {
            "description": "Method not allowed."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiErrorCode": {
        "type": "string",
        "enum": [
          "METHOD_NOT_ALLOWED",
          "AUTH_REQUIRED",
          "INVALID_JSON",
          "INVALID_CREDENTIALS_INPUT",
          "EMAIL_ALREADY_REGISTERED",
          "ACCOUNT_CREATE_FAILED",
          "INVALID_CREDENTIALS",
          "SESSION_SECRET_REQUIRED",
          "API_KEY_SECRET_REQUIRED",
          "INVALID_API_KEY",
          "API_KEY_SAVE_FAILED",
          "INVALID_CHAT_MESSAGE",
          "INVALID_CONVERSATION_ID",
          "CONVERSATION_NOT_FOUND",
          "API_KEY_REQUIRED",
          "API_KEY_DECRYPT_FAILED",
          "RUNTIME_REQUIRED",
          "CHAT_QUOTA_EXCEEDED",
          "UPSTREAM_CHAT_FAILED"
        ]
      },
      "SanitizedAuditEvent": {
        "type": "object",
        "required": ["type", "createdAt"],
        "properties": {
          "type": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          }
        }
      }
    }
  }
}